+++ packages/exception/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rym-lib/exception
 
+## 1.2.3
+
+### Patch Changes
+
+- update query-module
+
 ## 1.2.2
 
 ### Patch Changes