+++ packages/nakadachi-adapter-react-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rym-lib/nakadachi-adapter-react-router
 
+## 1.2.3
+
+### Patch Changes
+
+- update query-module
+- Updated dependencies
+  - @rym-lib/nakadachi@1.2.3
+
 ## 1.0.0
 
 ### Major Changes